configurando banco de dados

+++ api_web/src/main/java/api_web/api_web/model/Usuario/UsuarioDTO.java
@@ -12,7 +12,7 @@ public class UsuarioDTO {
     private String email;
     private String nome;
     private String descricao;
-    private long seguidores;
+    private int seguidores;
     private String senha;
